Fix SignIn, SignUp, SignOut

+++ admin-app/src/actions/auth.actions.js
@@ -23,41 +23,11 @@ export const login = (user) => {
                     token, user
                 }
             });
-        }else{
-            if(res.status === 400){
-                dispatch({
-                    tyoe: authConstants.LOGIN_FAILURE,
-                    payload: {error: res.data.error }
-                });
-            }
         }
-    }
-}
-
-export const signup = (user) => {
-
-    console.log(user);
-
-    return async (dispatch) => {
-
-        dispatch({ type:authConstants.LOGIN_REQUEST });
-
-        const res = await axios.post(`/admin/signup`, {
-             ...user
-        });
-
-        if(res.status === 201) {
-            const { message} = res.data;
-            dispatch({
-                type: authConstants.LOGIN_SUCCESS,
-                payload:{
-                    token, user
-                }
-            });
-        }else{
+        else{
             if(res.status === 400){
                 dispatch({
-                    tyoe: authConstants.LOGIN_FAILURE,
+                    type: authConstants.LOGIN_FAILURE,
                     payload: {error: res.data.error }
                 });
             }
@@ -78,7 +48,7 @@ export const isUserLoggedIn = () =>{
             });
         }else{
             dispatch({
-                tyoe: authConstants.LOGIN_FAILURE,
+                type: authConstants.LOGIN_FAILURE,
                 payload: {error: 'Failed to login' }
             });
         }
@@ -87,9 +57,17 @@ export const isUserLoggedIn = () =>{
 
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        });
+        dispatch({ type: authConstants.LOGOUT_REQUEST });
+        const res = await axios.post(`/admin/signout`);
+
+        if(res.status === 200){
+            localStorage.clear();
+            dispatch({ type: authConstants.LOGOUT_SUCCESS });
+        }else{
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE,
+                payload: { error: res.data.error }
+            });
+        }
     }
 }
